Add inline permission check

+++ test-project/api/controllers/things/destroy-one-thing.js
@@ -15,10 +15,25 @@ module.exports = {
   },
 
 
-  exits: {},
+  exits: {
+    forbidden: {
+      description: 'The user making this request is not permitted to delete this thing.',
+      responseType: 'forbidden',
+    }
+  },
 
 
   fn: async function (inputs, exits) {
+
+    let thing = await Thing.findOne({
+      id: inputs.id
+    });
+
+    console.log(thing);
+
+    if (thing.owner !== this.req.me.id) {
+      throw 'forbidden';
+    }
 
     await Thing.destroy({id: inputs.id});
